% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_assignment.R
\name{assign_overlaps_acc}
\alias{assign_overlaps_acc}
\title{Assign the overlap ID of a timestamp}
\usage{
assign_overlaps_acc(sample)
}
\description{
Assign the overlap ID of a timestamp
}
\examples{
#> Source Code:
assign_overlaps_acc
}
