% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_related.R
\name{back_convert_simulation}
\alias{back_convert_simulation}
\title{Convert Simmulation results back to data format}
\usage{
back_convert_simulation(data)
}
\description{
Convert Simmulation results back to data format
}
\examples{
#> Source Code:
back_convert_simmulation
}
