% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{bb}
\alias{bb}
\title{Define project bounding box}
\format{
An object of class \code{list} of length 2.
}
\usage{
bb
}
\description{
\code{bb} set the bounding box framing Isla Barro Colorado
}
\examples{
#> Source Code:
bb
}
\keyword{datasets}
