% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{bic_proj}
\alias{bic_proj}
\title{Project UTM crs}
\format{
An object of class \code{character} of length 1.
}
\usage{
bic_proj
}
\description{
\code{bic_proj} holds the project UTM projection.
}
\examples{
#> Source Code:
bic_proj
}
\keyword{datasets}
