% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{canal}
\alias{canal}
\title{Load Panama Canal}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 48 rows and 23 columns.
}
\usage{
canal
}
\description{
Load Panama Canal
}
\examples{
ggplot()+ geom_sf(data = canal)
}
\keyword{datasets}
