% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{cent_plotter}
\alias{cent_plotter}
\title{plot all centroids within a tibble-column of sf-objects}
\usage{
cent_plotter(sf_centroid, individual.local.identifier, ...)
}
\description{
plot all centroids within a tibble-column of sf-objects
}
\examples{
#> Source Code:
cent_plotter
}
