% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{central_point}
\alias{central_point}
\title{Central point (non-sfversion)}
\format{
An object of class \code{list} of length 2.
}
\usage{
central_point
}
\description{
Central point (non-sfversion)
}
\keyword{datasets}
