% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{central_point_crs}
\alias{central_point_crs}
\title{Set Projection of Central Point of Area Covered by Tracks}
\format{
An object of class \code{sfc_POINT} (inherits from \code{sfc}) of length 1.
}
\usage{
central_point_crs
}
\description{
Set Projection of Central Point of Area Covered by Tracks
}
\keyword{datasets}
