% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{central_point_sf}
\alias{central_point_sf}
\title{Get central point of area covered by tracks (as sf object)}
\format{
An object of class \code{XY} (inherits from \code{POINT}, \code{sfg}) of length 2.
}
\usage{
central_point_sf
}
\description{
Get central point of area covered by tracks (as sf object)
}
\examples{
#> Source Code:
central_point_sf
}
\keyword{datasets}
