% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{centroid_of_points}
\alias{centroid_of_points}
\title{combine several sf-points and create a joint centroid}
\usage{
centroid_of_points(x)
}
\description{
combine several sf-points and create a joint centroid
}
