% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{clr_accent}
\alias{clr_accent}
\title{Highlighter Color}
\format{
An object of class \code{character} of length 1.
}
\usage{
clr_accent
}
\description{
\code{clr_accent} defines a highlighter color.
}
\keyword{datasets}
