% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_colors.R
\docType{data}
\name{clr_neighbor}
\alias{clr_neighbor}
\title{The basic color scheme for pair_ids}
\format{
An object of class \code{colors} of length 10.
}
\usage{
clr_neighbor
}
\description{
\code{clr_neighbor} contains the base colors of the project color scheme for sample pairs
}
\keyword{datasets}
