% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_colors.R
\docType{data}
\name{clr_neighbor_alpha}
\alias{clr_neighbor_alpha}
\title{The semi-transparent color scheme for pair_ids}
\format{
An object of class \code{colors} of length 10.
}
\usage{
clr_neighbor_alpha
}
\description{
\code{clr_neighbor_alpha} contains a semi-transparent version of the project color scheme for sample pairs
}
\keyword{datasets}
