% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_colors.R
\docType{data}
\name{clr_set_alpha}
\alias{clr_set_alpha}
\title{The transparent project color scheme}
\format{
An object of class \code{colors} of length 8.
}
\usage{
clr_set_alpha
}
\description{
The transparent project color scheme
}
\examples{
clr_set_alpha
}
\seealso{
[fftidy::clr_set_base] a basic version of the color scheme.

[fftidy::clr_set_samples] a less saturated version of the color scheme.

[fftidy::clr_set_light] a lighter version of the color scheme.
}
\keyword{datasets}
