% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_colors.R
\docType{data}
\name{clr_set_base}
\alias{clr_set_base}
\title{The basic color scheme for the sample IDs}
\format{
An object of class \code{colors} of length 8.
}
\usage{
clr_set_base
}
\description{
\code{clr_set_base} contains base colors of the project color scheme
}
\details{
The color scheme is used to achieve a consistent usage of color throughout the project.
}
\examples{
#> Source Code:
clr_set_base
clr_set_samples
clr_set_light
clr_set_alpha

names(clr_set_base)
}
\seealso{
[fftidy::clr_set_samples] a less saturated version of the color scheme.

[fftidy::clr_set_light] a lighter version of the color scheme.

[fftidy::clr_set_alpha] a transparent version of the color scheme.
}
\keyword{datasets}
