% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{clr_water}
\alias{clr_water}
\title{Water Color}
\format{
An object of class \code{colors} of length 1.
}
\usage{
clr_water
}
\description{
\code{clr_water} defines a water color for maps.
}
\keyword{datasets}
