% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{compare_all_models}
\alias{compare_all_models}
\title{function to compare variogram with all ML model fits (initial close up and total data range)}
\usage{
compare_all_models(id, svfs, fitted_mods, ...)
}
\description{
function to compare variogram with all ML model fits (initial close up and total data range)
}
