% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{compare_n_best_models}
\alias{compare_n_best_models}
\title{function to compare variogram with the n best ML model fits (initial close up and total data range)}
\usage{
compare_n_best_models(id, svfs, n_mod_max = 5, fitted_mods, ...)
}
\description{
function to compare variogram with the n best ML model fits (initial close up and total data range)
}
