% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighboring_groups.R
\name{complex_distance}
\alias{complex_distance}
\title{Compute pair wise distances within time bins}
\usage{
complex_distance(data, ind1, ind2, ...)
}
\description{
Compute pair wise distances within time bins based on complex representation
}
