% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_extend_data.R
\name{data_extend}
\alias{data_extend}
\title{Extend tracking data with computed metrics}
\usage{
data_extend(data)
}
\arguments{
\item{data}{the tracking data (imported using \code{data_fft()})}
}
\description{
\code{extend_data} computes additional simple metrics from the raw capuchin tracking data
}
\examples{
ext_data <- data_fft(data_path = "data/fft_capuchins_only.csv.gz") \%>\% data_extend()
head(ext_data)
}
\seealso{
[fftidy::extend_data()] a less saturated version of the color scheme.
}
