% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_load_data.R
\name{data_fft}
\alias{data_fft}
\title{Load FFT tracking data}
\usage{
data_fft(data_path = "data/fft_capuchins_movebank_download.tsv.gz")
}
\arguments{
\item{data_path}{the path to the data file}
}
\description{
\code{data_fft} imports the capuchin tracking data.
}
\details{
By default, the function expects to be run from the FFT project folder .
}
\examples{
#> Source Code:
data_fft
}
\seealso{
[fftidy::extend_data()] a less saturated version of the color scheme.
}
