% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{df_as_sf}
\alias{df_as_sf}
\title{create the convex hull of a two column tibble with lat long columns (output as sf-obj)}
\usage{
df_as_sf(x)
}
\description{
create the convex hull of a two column tibble with lat long columns (output as sf-obj)
}
\examples{
#> Source Code:
df_as_sf
}
