% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{dipteryx}
\alias{dipteryx}
\title{Load Locations of Dipteryx}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1225 rows and 5 columns.
}
\usage{
dipteryx
}
\description{
Load Locations of Dipteryx
}
\keyword{datasets}
