% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{dtim}
\alias{dtim}
\title{extract the time of the day from a date-time (as numeric seconds)}
\usage{
dtim(x)
}
\description{
extract the time of the day from a date-time (as numeric seconds)
}
\examples{
#> Source Code:
dtim
}
