% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_assignment.R
\name{encounter_long}
\alias{encounter_long}
\title{Pivot encointer table to long format}
\usage{
encounter_long(data, sample_id_n = "sample_id_e")
}
\description{
Pivot encointer table to long format
}
\examples{
#> Source Code:
encounter_long
}
