% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{fft_case}
\alias{fft_case}
\title{Project case}
\usage{
fft_case(x)
}
\description{
\code{fft_case} manages the format of labels for figure panels.
}
\details{
Currently all figure sub panels are labelled using lower case.
This function potentially needs to be changed in case of
resubmission to a different journal.
}
\examples{
fft_case(letters[1:5])
fft_case(LETTERS[1:5])
}
