% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{fft_case_inv}
\alias{fft_case_inv}
\title{Inverted project case}
\usage{
fft_case_inv(x)
}
\description{
\code{fft_case_inv} manages the format of labels within figures.
}
\details{
Labels within figures (outlier highlights) should be in inverse case
compared to figure panel label to avoid confusion.
This function potentially needs to be changed in case of
resubmission to a different journal.
}
\examples{
#> Source Code:
fft_case_inv(letters[1:5])
fft_case_inv(LETTERS[1:5])
}
