% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{fft_f_width}
\alias{fft_f_width}
\title{Default figure width (double column)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
fft_f_width
}
\description{
\code{fft_f_width} defines the full width for figures (mm).
}
\details{
7 inch = 178mm
}
\examples{
fft_f_width
}
\keyword{datasets}
