% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{fft_f_width_half}
\alias{fft_f_width_half}
\title{Default figure width (single column)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
fft_f_width_half
}
\description{
\code{fft_f_width_half} defines the half width for figures (mm).
}
\details{
3.43 inch = 87mm
}
\examples{
fft_f_width_half
}
\keyword{datasets}
