% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{fft_lwd}
\alias{fft_lwd}
\title{Default plot linewidth}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
fft_lwd
}
\description{
\code{fft_lwd} sets the linewidth in plots.
}
\details{
This sets the default linewidth thoughout the figures.
}
\examples{
fft_lwd
}
\keyword{datasets}
