% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{fft_plot_size}
\alias{fft_plot_size}
\title{Default plot size}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
fft_plot_size
}
\description{
\code{fft_plot_size} sets the default size of data points in plots.
}
\details{
This sets the default size of data points thoughout the figures.
}
\keyword{datasets}
