% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{fft_text_size}
\alias{fft_text_size}
\title{Default plot font size}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
fft_text_size
}
\description{
\code{fft_text_size} sets the default font size of text in plots.
}
\details{
This sets the default font size of text thoughout the figures.
}
\examples{
#> Source Code:
fft_text_size
}
\keyword{datasets}
