% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{fft_text_size_small}
\alias{fft_text_size_small}
\title{Small plot font size}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
fft_text_size_small
}
\description{
\code{fft_text_size_small} sets the small font size of text in plots.
}
\details{
This sets the small font size of text for the figures (legends).
}
\examples{
#> Source Code:
fft_text_size_small
}
\keyword{datasets}
