% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{fit_ouf}
\alias{fit_ouf}
\title{function to create and fit a Ornstein-Uhlenbeck-F model to a semi-variance function}
\usage{
fit_ouf(svf, sigma, tauH, tauF)
}
\description{
function to create and fit a Ornstein-Uhlenbeck-F model to a semi-variance function
}
\examples{
#> Source Code:
fit_ouf
}
