% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{gatun}
\alias{gatun}
\title{Load Gatun Lake Outline}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
gatun
}
\description{
Load Gatun Lake Outline
}
\examples{
#> Source Code:
gatun
#> -------------------
ggplot()+ geom_sf(data = gatun)
}
\keyword{datasets}
