% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_akde}
\alias{get_akde}
\title{function for automatizing autocorrelated kernel density estimation}
\usage{
get_akde(data, fitted_mls, id, ...)
}
\description{
function for automatizing autocorrelated kernel density estimation
}
\examples{
#> Source Code:
get_akde
}
