% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_akde_aligned}
\alias{get_akde_aligned}
\title{Load ctmm based AKDEs grouped by field season}
\usage{
get_akde_aligned(data, fitted_mls, year, ...)
}
\description{
Load ctmm based AKDEs grouped by field season
}
