% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_akde_maual}
\alias{get_akde_maual}
\title{Load ctmm based AKDEs by Sample ID}
\usage{
get_akde_maual(data, fitted_mls, id, ...)
}
\description{
Load ctmm based AKDEs by Sample ID
}
\examples{
#> Source Code:
get_akde_maual
}
