% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_ci_ouf}
\alias{get_ci_ouf}
\title{function for OUF specific parameters (tau2)}
\usage{
get_ci_ouf(mod, ci_idx, ci_type = "est", ...)
}
\description{
function for OUF specific parameters (tau2)
}
