% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_ci_univ}
\alias{get_ci_univ}
\title{universal function (works for both OU & OUF)}
\usage{
get_ci_univ(mod, ci_idx, ci_type = "est", ...)
}
\description{
universal function (works for both OU & OUF)
}
\examples{
#> Source Code:
get_ci_univ
}
