% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_kriged_occurrence_maual}
\alias{get_kriged_occurrence_maual}
\title{Load ctmm based kriged occurrences by Sample ID}
\usage{
get_kriged_occurrence_maual(data, fitted_mls, id, ...)
}
\description{
Load ctmm based kriged occurrences by Sample ID
}
