% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_param_estimate}
\alias{get_param_estimate}
\title{Get the parameter estimates from a ctmm fitted model}
\usage{
get_param_estimate(mod)
}
\description{
Get the parameter estimates from a ctmm fitted model
}
\examples{
#> Source Code:
get_param_estimate
}
