% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{get_pw_dists}
\alias{get_pw_dists}
\title{given a table with synchronous positions, compute all distances for a specific column-pair (sample-pair)}
\usage{
get_pw_dists(data, ind1, ind2, ...)
}
\description{
given a table with synchronous positions, compute all distances for a specific column-pair (sample-pair)
}
\examples{
#> Source Code:
get_pw_dists
}
