% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_var_fit_ou}
\alias{get_var_fit_ou}
\title{Extract a OU fitted parameter from a ctmm object}
\usage{
get_var_fit_ou(vario, xlim = c(0, 50), n = 200)
}
\description{
Extract a OU fitted parameter from a ctmm object
}
\examples{
#> Source Code:
get_var_fit_ou
}
