% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_var_fit_ouf}
\alias{get_var_fit_ouf}
\title{Extract a OUF fitted parameter from a ctmm object}
\usage{
get_var_fit_ouf(vario, xlim = c(0, 50), n = 200)
}
\description{
Extract a OUF fitted parameter from a ctmm object
}
