% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{get_var_fit_raw}
\alias{get_var_fit_raw}
\title{Extract a fitted parameter from a ctmm object}
\usage{
get_var_fit_raw(vario, model = "ou")
}
\description{
Extract a fitted parameter from a ctmm object
}
