% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{hull_tib}
\alias{hull_tib}
\title{create the convex hull of a two column tibble with lat long columns (output as tibble)}
\usage{
hull_tib(x)
}
\description{
create the convex hull of a two column tibble with lat long columns (output as tibble)
}
\examples{
#> Source Code:
hull_tib
}
