% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{idx_graph}
\alias{idx_graph}
\title{convert an igraph-graph into a tidygraph tbl_graph, annotate with the idx (later to become the time-bin), and compute tdiverse node-centralities}
\usage{
idx_graph(g, idx)
}
\description{
convert an igraph-graph into a tidygraph tbl_graph, annotate with the idx (later to become the time-bin), and compute tdiverse node-centralities
}
\examples{
#> Source Code:
idx_graph
}
