% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{island}
\alias{island}
\title{Load Isla Barro Colorado Outline}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
island
}
\description{
Load Isla Barro Colorado Outline
}
\keyword{datasets}
