% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{merge_sf_columns}
\alias{merge_sf_columns}
\title{merge multiple sf columns}
\usage{
merge_sf_columns(col, data)
}
\description{
merge multiple sf columns
}
\examples{
#> Source Code:
merge_sf_columns
}
