% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{mod_tab}
\alias{mod_tab}
\title{wrapper to get sigma, tauH, tauF, and speed estimates of a model}
\usage{
mod_tab(mod)
}
\description{
wrapper to get sigma, tauH, tauF, and speed estimates of a model
}
