% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{neighbor_ids}
\alias{neighbor_ids}
\title{Vector of all neighbor-pair_ids}
\format{
An object of class \code{character} of length 10.
}
\usage{
neighbor_ids
}
\description{
\code{neighbor_ids} contains all five character neighbor pair_ids.
}
\examples{
#> Source Code:
neighbor_ids
}
\keyword{datasets}
