% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighboring_groups.R
\docType{data}
\name{neighboring_samples}
\alias{neighboring_samples}
\title{Neighboring groups}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
neighboring_samples
}
\description{
Groups neighboring each other (possible encouters)
}
\keyword{datasets}
