% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouring_groups.R
\docType{data}
\name{neighbouring_samples}
\alias{neighbouring_samples}
\title{Neigbouring groups}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
neighbouring_samples
}
\description{
Groups neighbouring each other (possible encouters)
}
\keyword{datasets}
