% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{outer_hull}
\alias{outer_hull}
\title{Get area covered by tracks}
\format{
An object of class \code{XY} (inherits from \code{POLYGON}, \code{sfg}) of length 1.
}
\usage{
outer_hull
}
\description{
\code{outer_hull} creates the outer hull of all the track within the data set.
}
\examples{
#> Source Code:
outer_hull
}
\keyword{datasets}
