% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{outer_hull_sf}
\alias{outer_hull_sf}
\title{sf version of area covered by tracks}
\format{
An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.
}
\usage{
outer_hull_sf
}
\description{
sf version of area covered by tracks
}
\examples{
#> Source Code:
outer_hull_sf
}
\keyword{datasets}
