% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{p_init_hull}
\alias{p_init_hull}
\title{initialize a field-trip specific base-plot (with individuals range)}
\usage{
p_init_hull(season)
}
\description{
initialize a field-trip specific base-plot (with individuals range)
}
