% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{p_init_voronoi}
\alias{p_init_voronoi}
\title{Base plot for voronoi layers}
\usage{
p_init_voronoi(season)
}
\description{
Base plot for voronoi layers
}
\examples{
#> Source Code:
p_init_voronoi
}
