% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{pan_detail}
\alias{pan_detail}
\title{Load clipped Panama Coastline}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
pan_detail
}
\description{
Load clipped Panama Coastline
}
\keyword{datasets}
