% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{parse_cis}
\alias{parse_cis}
\title{functions to parse estimated model parameters (mu & tau) from ctmm}
\usage{
parse_cis(mod, ci_type)
}
\description{
get the estimate units and round estimate to two digits
}
\examples{
#> Source Code:
parse_cis
}
