% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{plot_ctm}
\alias{plot_ctm}
\title{plot two model (original guess and modified) in context of variogram (in right colors)}
\usage{
plot_ctm(mod, svf, ...)
}
\description{
plot two model (original guess and modified) in context of variogram (in right colors)
}
