% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{plot_svf}
\alias{plot_svf}
\title{modified plot.svf from {ctmm}}
\usage{
plot_svf(lag, CTMM, error = NULL, alpha = 0.05, col = "red", type = "l", ...)
}
\description{
The function was altered from ctmm:::plot.svf()
}
\details{
It now returns the internally computed SVF
}
\examples{
#> Source Code:
plot_svf
}
