% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{plot_svf_vs_model}
\alias{plot_svf_vs_model}
\title{function to compare variogram with a model fit (initial close up and total data range)}
\usage{
plot_svf_vs_model(variogram, mod, colors = clr_set_base, ...)
}
\description{
function to compare variogram with a model fit (initial close up and total data range)
}
