% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{plot_svfs}
\alias{plot_svfs}
\title{function to create a variogram plot pair (initial close up and total data range)}
\usage{
plot_svfs(variogram, colors = clr_set_base)
}
\description{
function to create a variogram plot pair (initial close up and total data range)
}
\examples{
#> Source Code:
plot_svfs
}
