% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{plot_svfs_single}
\alias{plot_svfs_single}
\title{function to plot a variogram in the correct color of athe sample}
\usage{
plot_svfs_single(variogram, colors = clr_set_base)
}
\description{
function to plot a variogram in the correct color of athe sample
}
