% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{sample_ids}
\alias{sample_ids}
\title{The sample IDs}
\format{
An object of class \code{character} of length 8.
}
\usage{
sample_ids
}
\description{
\code{sample_ids} contains the sample name and the collar ID.
}
\details{
The IDs are ordered according to their occurrence in the FFT data set.
}
\keyword{datasets}
