% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{sample_ids_short}
\alias{sample_ids_short}
\title{Abbreviations of sample IDs}
\format{
An object of class \code{character} of length 8.
}
\usage{
sample_ids_short
}
\description{
\code{sample_ids_short} contains a two character sample ID abbreviation.
}
\keyword{datasets}
