% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_colors.R
\docType{data}
\name{sample_years}
\alias{sample_years}
\title{The sample year-class}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
sample_years
}
\description{
\code{sample_years} contains the field season for each sample.
}
\examples{
sample_years
}
\keyword{datasets}
