% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{samples_short}
\alias{samples_short}
\title{Reference vector short to long sample ID}
\format{
An object of class \code{character} of length 8.
}
\usage{
samples_short
}
\description{
\code{samples_short} contains a named vector of the sample IDs.
}
\keyword{datasets}
