% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{separate_mod_tab}
\alias{separate_mod_tab}
\title{separate  ctmm model estimates from units}
\usage{
separate_mod_tab(tab)
}
\description{
separate  ctmm model estimates from units
}
\examples{
#> Source Code:
separate_mod_tab
}
