% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\name{sf_raster_empty}
\alias{sf_raster_empty}
\title{Initialize empty Stars object}
\usage{
sf_raster_empty(sf, res = 50)
}
\description{
Initialize empty Stars object
}
\examples{
#> Source Code:
sf_raster_empty
}
