% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_related.R
\name{sim_to_sf}
\alias{sim_to_sf}
\title{Convert simulated tracks into a sf-object}
\usage{
sim_to_sf(SIM, id_nr, type)
}
\description{
used in simulate_id()
}
\examples{
#> Source Code:
sim_to_sf
}
