% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_related.R
\name{simulate_id}
\alias{simulate_id}
\title{Simulate movement track for a particular Sample}
\usage{
simulate_id(data, id_nr, ctmm_selcetion_type, fitted_tib)
}
\description{
Simulate movement track for a particular Sample
}
\examples{
#> Source Code:
simulate_id
}
