% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{subs}
\alias{subs}
\title{subset a list of lists by a common index (the n_th object of each sub-list)}
\usage{
subs(x, idx)
}
\description{
subset a list of lists by a common index (the n_th object of each sub-list)
}
\examples{
#> Source Code:
subs
}
