% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.R
\docType{data}
\name{table_theme}
\alias{table_theme}
\title{Basic table theme}
\format{
An object of class \code{reactableTheme} of length 1.
}
\usage{
table_theme
}
\description{
Basic table theme
}
\examples{
#> Source Code:
table_theme
}
\keyword{datasets}
