% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R, R/geo_related.R
\name{tibble_to_sf}
\alias{tibble_to_sf}
\title{turn a tibble with a lat and long column into a set sf points}
\usage{
tibble_to_sf(tib, crs = 4326)

tibble_to_sf(tib, crs = 4326)
}
\description{
turn a tibble with a lat and long column into a set sf points

Convert GPS tibble into sf object
}
\examples{
#> Source Code:
tibble_to_sf
#> Source Code:
tibble_to_sf
}
