% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{update_model}
\alias{update_model}
\title{manipulate ctmm model and output new model alongside original for comparison plot}
\usage{
update_model(mod, update_fun, ...)
}
\description{
manipulate ctmm model and output new model alongside original for comparison plot
}
