% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_functions.R
\name{update_tau}
\alias{update_tau}
\title{manipulate ctmm model tau entry}
\usage{
update_tau(mod, tauH, tauF, ...)
}
\description{
manipulate ctmm model tau entry
}
\examples{
#> Source Code:
update_tau
}
