% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_related.R
\docType{data}
\name{wgs}
\alias{wgs}
\title{The WGS proj string}
\format{
An object of class \code{character} of length 1.
}
\usage{
wgs
}
\description{
\code{wgs} contains the proj string for GPS data
}
\examples{
wgs
}
\keyword{datasets}
