% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fft_functions.R
\name{wide_to_distmat}
\alias{wide_to_distmat}
\title{transform a wide tibble with pair-wise distances into a distance-matrix}
\usage{
wide_to_distmat(tib)
}
\description{
transform a wide tibble with pair-wise distances into a distance-matrix
}
\examples{
#> Source Code:
wide_to_distmat
}
